Split traits out of ExtractionUtils and into their own files
